+++ cli/lib/Cargo.toml
@@ -20,6 +20,7 @@ deno_fs = { workspace = true, features = ["sync_fs"] }
 deno_media_type.workspace = true
 deno_node = { workspace = true, features = ["sync_fs"] }
 deno_npm.workspace = true
+deno_npm_installer.workspace = true
 deno_path_util.workspace = true
 deno_resolver = { workspace = true, features = ["sync"] }
 deno_runtime.workspace = true
